Crate axvmconfig

ArceOS-Hypervisor VM config module. AxVMCrateConfig: the configuration structure for the VM. It is generated from toml file, and then converted to AxVMConfig for the VM creation.

Structs

AxVMCrateConfig

The configuration structure for the guest VM serialized from a toml file provided by user, and then converted to AxVMConfig for the VM creation.

EmulatedDeviceConfig

A part of AxVMConfig, which represents the configuration of an emulated device for a virtual machine.

PassThroughDeviceConfig

A part of AxVMConfig, which represents the configuration of a pass-through device for a virtual machine.

VmMemConfig

A part of AxVMConfig, which represents a memory region.

Enums

VMType

A part of AxVMConfig, which represents guest VM type.

VmMemMappingType

The type of memory mapping.